requirements.txt
python-dotenv>=1.0.0
PyYAML>=6.0
gspread>=6.0.0
oauth2client>=4.1.3
google-auth>=2.0.0
google-auth-oauthlib>=0.5.3
google-api-python-client>=2.100.0
supabase>=1.0.3